Modified  sieving in APLOD.
Now, sieving will only be used if numComponents == 1 (in this case, only one seek is required).
Otherwise, the full PG will be read, but only the components specified for the query
(thus, still only one seek is used). This is because it seems seeks are quite expensive, so we
limit them in exchange for a bit of extra data read. In the future, a more complex heuristic
could be used.