Fixed issue trying to set CMAKE_INSTALL_PREFIX.
The only way to set the install directory for ADIOS was by creating
an INSTALL_PREFIX environment variable. The CMake variable would get cleared out
if that value wasn't set. This change checks first to see if CMAKE_INSTALL_PREFIX
is set before trying to get the environment variable value to set as
CMAKE_INSTALL_PREFIX.