serial adios_open with parallel lib
Fix a runtime error in the following situation:

when a user compiles a serial program against the parallel
ADIOS libs, there are two ways to use ADIOS without the need
to start the app with `mpiexec`:

a) use the moc library instead
b) pass `MPI_COMM_NULL` as communicator

I need the last case in my application but still `adios_open`
calls a single `MPI_Comm_rank` which aborts the application.

This check (as it is done in many other places) fixes it.