Numpy: Scalar Attributes Warning
Upgrade the conversion of returned scalar attributes
to the recommended usage in numpy.

In Numpy 1.16 and newer, the old usage throws a warning of
the form
```
DeprecationWarning: np.asscalar(a) is deprecated since NumPy v1.16, use a.item() instead
  This is separate from the ipykernel package so we can avoid doing imports until
```

It is likely this will be removed in Numpy 1.18 and newer, so
let us fix this right away.

`ndarray.item()` is available since a long time, so that is a save
change.